[REFACTOR] notice util로 분리 #7
