+++ src/pages/NoticeListPage.vue
@@ -103,6 +103,7 @@ import Pagination from '@/components/ui/pagination/Pagination.vue'
 import LoginAlertModal from '@/components/login/LoginAlertModal.vue'
 import NoticeFilter from '@/components/searchNotice/NoticeFilter.vue'
 import { formatDate } from '@/utils/dateUtils'
+import { applyStatus, statusColorClass } from '/src/utils/noticeUtils.js'
 
 const router = useRouter()
 
@@ -162,27 +163,6 @@ const fetchNotices = async () => {
   } finally {
     loading.value = false
   }
-}
-
-const applyStatus = (notice) => {
-  const begin = new Date(notice.beginDate)
-  const end = new Date(notice.endDate)
-  if (!notice.beginDate || !notice.endDate) return '정보없음'
-  if (today < begin) return '모집예정'
-  if (today >= begin && today <= end) return '모집중'
-  if (today > end) return '모집완료'
-  return '정보없음'
-}
-
-const statusColorClass = (notice) => {
-  const status = applyStatus(notice)
-  const colorMap = {
-    '모집예정': 'bg-yellow-100 text-yellow-800',
-    '모집중': 'bg-red-100 text-red-800',
-    '모집완료': 'bg-green-100 text-green-800',
-    '정보없음': 'bg-gray-100 text-gray-600'
-  }
-  return colorMap[status] || 'bg-gray-100 text-gray-600'
 }
 
 const supplyTypeColorClass = (type) => {
